fix: 🐛 int coordinates

+++ common.py
@@ -1,3 +1,3 @@
 """Tipos de datos comunes para el proyecto."""
 
-Coordinate = tuple[float, float]
+Coordinate = tuple[int, int]
